First commit to the repository
While a mockup rather than a fully working implementation, this
change shows what the app will look like in basic states when it is
fully implemented. It has two routes: the home/default route and the
route for creating a new note, or editing an existing note.

Changes to the mock up for deleting a note and for when a form fails
validation may come later.