Update the UX
This change updates the mockup in a number of key ways. It updates the
list of notes on the home page to use the same edit/delete button code.
It adds in two dropdowns above the list of notes; one for filtering
by category and one for sorting available records. It adds in error
formatting, for when errors are present validating the create note form.
In addition, there are a number of other, smaller, changes to refine the
quality of the UX.