Implement an early PHP version of the application
This change adds just about the most minimal version of the application
using Mini Mezzio that's possible. It provides a minimal
public/index.php which sets up the DI container and provides all of the
application's routes/endpoints. Only the default route and create note
route work, but it's a start.