Create a small database service
This change adds a small database service class to the project, based on
Axleus-Db's TableGateway class. It's a little unstructured, but provides
a quick and easy way to quickly get up and running with database
connectivity. In future, perhaps, the class will be refined to be more
structured, but for the time being, it works well enough as a rapid PoC.