Add laminas-inputfilter support
This change updates the project's PHP dependencies so that the draft PR
adding support for laminas-servicemanager v4 can be added to the project.
I tried a myriad of combinations to get laminas-servicemanager, mini
mezzio, laminas-inputfilter, and laminas-db/php-db to work, without
success, until I tried this one. It's a little hairy, but will become
less so, once the draft PR is merged and a new release of
laminas-inputfilter is pushed out.

The change also removes doctrine-orm support, which was never required.