Update backend dependencies
This change adds in a series of new backend dependencies required to support upcoming functionality.

Specifically, it adds in league/commonmark, league/html-to-markdown and twig/markdown-extra to support using Twig’s markdown_to_html filter in the view-notes template.

And, it adds php-db/phpdb-adapter-sqlite to add full support for working with SQLite databases. php-db/phpdb was already available, but I’d forgotten that this package does not have support for specific database vendors, rather it contains the code common to all database vendors supported by the project.