Render a note's body as HTML when viewing notes
After a bit of thought on this, effectively storing a note's body as
Markdown (I'd prefer to store it as AsciiDoc) solves a couple of small
issues with rendering it post-creation/edit. The primary one is that new
lines don't need to be converted to HTML breaks, etc. The text can just
be left as is and then rendered as HTML when viewed.

Given that, this change passes the note's body through the Twig
markdown_to_html filter, rendering it as HTML.