Fix up the implementation of the manage note template's submit button
This change refactors the button so that it properly reflects the state
in effect at the time. By that, I mean that if the template is for
creating a new note, then the submit button will be a "create" button.
However, if the template is for updating an existing note, then the
submit button will be an "update" button. In both contexts, the
appropriate theme (colours and icon) will be applied.