Reflect creating and editing notes properly
The manage notes template, up until this point, really was the create note template, with some changes to reflect updating notes available. Given that the template is used for both creating and updating notes, this could not continue.

So, in this change it has been updated to ensure that no matter which operation it is being used for, it fully and properly reflects it.