Add note that the body field supports Markdown formatting
This follows the recent change to effectively add Markdown support in a note’s body content. It adds a notice/tip/hint underneath the body form in the manage note form telling the user that Markdown formatting is supported in the body field, with a link to markdownguide.org, arguably the best Markdown guide on the internet, imho.

Add note about Markdown addition