Add and enable an InputFilter for notes
To provide self-contained input validation and filtering logic for notes, this changes introduces a custom InputFilter object. The change also adds in a number of pre-existing ConfigProviders that enable the relevant, underlying dependencies so that the functionality will work smoothly.

As a side note, in Mezzio, the ConfigProviders would have been enabled when the packages were enabled, leading me, again, to consider going back to Mezzio. However, for the time being, I’m continuing a bit longer with Mini Mezzio.