Implement more of the create and edit functionality
This change continues implementing the note create and edit functionality, getting it most of the way to completion. Specifically, notes can now be edited and created, performing redirects when required either after a successful creation or update, or when the form does not validate.

What’s missing, however, is any notification to the user when a form doesn’t validate or when an operation was successful. This is coming in future changes.