Integrate the rollup.js module bundler
This change provides the simplest integration of rollup.js into the project to bundle the app’s JavaScript file.

This avoids the need to add a potentially long list of CDN links to the default template, helps keep the public directory free from any file that can be built at build time, and helps make the code easier to maintain over the course of time.

The configuration includes the nodeResolve plugin to include packages installed with npm.