Reorganise the application’s stylesheet
This change moves the application’s non-generated stylesheet to a more logical and intuitive directory in the src/Assets directory.

It’s not essential to be done, especially not now, but it makes sense to do it now in preparation for later, and because I’m aware that it’s something that should be done, so why not do it.