Refactor the delete endpoint to be a POST request
After thinking through how the app works, again, I realised that perhaps the most logical way is to use a form, perhaps as I did previously, to submit a request to delete a note.

By doing that, it’s easier for JavaScript to set the note’s id in the form in the delete dialog, and it’s perhaps a bit easier when processing the delete request.