Final styling changes
This change makes a series of small changes to the styling of the
application. It fixes a small padding issue with pills. It fixes a small
styling change to links within a note's body. It removes empty styling
for the confirmation dialog. It updates the dialog's styling to be layed
out properly horizontally on the page. Finally, it removes excess bottom
padding for h1 headers.