Optimise Twig templates
This change does two things.

First, it moves some template aspects to includable templates, such as
the notifications and confirmation dialog. The reason for this is to
have each template be as focused as possible on as little or as focused
a task as possible, and to minimise code duplication, which was
happening with the notifications.

Second, it makes use of the ternary operator, where possible to have the
templates be more compact, yet also easier to read. While it was fine
before, this new approach, to me, is more maintainable, and quicker to
get your head across.