Add a static route for the about page
This change adds the route's static template file, adds styling for it,
and a route as well. The template makes use of the Twig markdown_to_html
filter to simplify creation of the page's content. I find it far easier
to create content rapidly this way, rather than manually crafting HTML.